+++ engine/nodejs_init.go
@@ -2,7 +2,6 @@ package engine
 
 import (
 	_ "github.com/dop251/goja_nodejs/buffer"
-	_ "github.com/dop251/goja_nodejs/process"
 	_ "github.com/dop251/goja_nodejs/url"
 	_ "github.com/dop251/goja_nodejs/util"
 )